backend/pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = ""
authors = ["Mpcsj"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
openai = "^0.27.6"
langchain = "^0.0.157"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
boto3 = "^1.26.126"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

